+++ androidx-kt/src/main/java/me/panpf/androidxkt/content/Contextx.kt
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+@file:Suppress("NOTHING_TO_INLINE")
+
 package me.panpf.androidxkt.content
 
 import android.accounts.AccountManager
@@ -78,195 +80,195 @@ import me.panpf.androidx.os.storage.StorageManagerCompat
 /* ******************************************* SystemService *******************************************/
 
 
-fun <T> Context.systemService(serviceName: String): T = Contextx.systemService(this, serviceName)
+inline fun  <T> Context.systemService(serviceName: String): T = Contextx.systemService(this, serviceName)
 
-fun <T> Context.systemServiceOrNull(serviceName: String): T? = Contextx.systemServiceOrNull(this, serviceName)
+inline fun  <T> Context.systemServiceOrNull(serviceName: String): T? = Contextx.systemServiceOrNull(this, serviceName)
 
-fun Context.packageManager(): PackageManager = Contextx.packageManager(this)
+inline fun  Context.packageManager(): PackageManager = Contextx.packageManager(this)
 
-fun Context.powerManager(): PowerManager = Contextx.powerManager(this)
+inline fun  Context.powerManager(): PowerManager = Contextx.powerManager(this)
 
-fun Context.windowManager(): WindowManager = Contextx.windowManager(this)
+inline fun  Context.windowManager(): WindowManager = Contextx.windowManager(this)
 
-fun Context.layoutInflater(): LayoutInflater = Contextx.layoutInflater(this)
+inline fun  Context.layoutInflater(): LayoutInflater = Contextx.layoutInflater(this)
 
-fun Context.accountManager(): AccountManager = Contextx.accountManager(this)
+inline fun  Context.accountManager(): AccountManager = Contextx.accountManager(this)
 
-fun Context.activityManager(): ActivityManager = Contextx.activityManager(this)
+inline fun  Context.activityManager(): ActivityManager = Contextx.activityManager(this)
 
-fun Context.alarmManager(): AlarmManager = Contextx.alarmManager(this)
+inline fun  Context.alarmManager(): AlarmManager = Contextx.alarmManager(this)
 
-fun Context.notificationManager(): NotificationManager = Contextx.notificationManager(this)
+inline fun  Context.notificationManager(): NotificationManager = Contextx.notificationManager(this)
 
-fun Context.accessibilityManager(): AccessibilityManager = Contextx.accessibilityManager(this)
+inline fun  Context.accessibilityManager(): AccessibilityManager = Contextx.accessibilityManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-fun Context.captioningManager(): CaptioningManager = Contextx.captioningManager(this)
+inline fun  Context.captioningManager(): CaptioningManager = Contextx.captioningManager(this)
 
-fun Context.keyguardManager(): KeyguardManager = Contextx.keyguardManager(this)
+inline fun  Context.keyguardManager(): KeyguardManager = Contextx.keyguardManager(this)
 
-fun Context.locationManager(): LocationManager = Contextx.locationManager(this)
+inline fun  Context.locationManager(): LocationManager = Contextx.locationManager(this)
 
-fun Context.searchManager(): SearchManager = Contextx.searchManager(this)
+inline fun  Context.searchManager(): SearchManager = Contextx.searchManager(this)
 
-fun Context.sensorManager(): SensorManager = Contextx.sensorManager(this)
+inline fun  Context.sensorManager(): SensorManager = Contextx.sensorManager(this)
 
-fun Context.storageManager(): StorageManager = Contextx.storageManager(this)
+inline fun  Context.storageManager(): StorageManager = Contextx.storageManager(this)
 
-fun Context.storageManagerCompat(): StorageManagerCompat = Contextx.storageManagerCompat(this)
+inline fun  Context.storageManagerCompat(): StorageManagerCompat = Contextx.storageManagerCompat(this)
 
 @RequiresApi(api = Build.VERSION_CODES.O)
-fun Context.storageStatsManager(): StorageStatsManager = Contextx.storageStatsManager(this)
+inline fun  Context.storageStatsManager(): StorageStatsManager = Contextx.storageStatsManager(this)
 
-fun Context.wallpaperManager(): WallpaperManager = Contextx.wallpaperManager(this)
+inline fun  Context.wallpaperManager(): WallpaperManager = Contextx.wallpaperManager(this)
 
 // TIME_ZONE_RULES_MANAGER_SERVICE
 
-fun Context.vibrator(): Vibrator = Contextx.vibrator(this)
+inline fun  Context.vibrator(): Vibrator = Contextx.vibrator(this)
 
-fun Context.connectivityManager(): ConnectivityManager = Contextx.connectivityManager(this)
+inline fun  Context.connectivityManager(): ConnectivityManager = Contextx.connectivityManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.P)
-fun Context.ipSecManager(): IpSecManager = Contextx.ipSecManager(this)
+inline fun  Context.ipSecManager(): IpSecManager = Contextx.ipSecManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.M)
-fun Context.networkStatsManager(): NetworkStatsManager = Contextx.networkStatsManager(this)
+inline fun  Context.networkStatsManager(): NetworkStatsManager = Contextx.networkStatsManager(this)
 
-fun Context.wifiManager(): WifiManager = Contextx.wifiManager(this)
+inline fun  Context.wifiManager(): WifiManager = Contextx.wifiManager(this)
 
-fun Context.wifiManagerOrNull(): WifiManager? = Contextx.wifiManagerOrNull(this)
-
-@RequiresApi(api = Build.VERSION_CODES.O)
-fun Context.wifiAwareManager(): WifiAwareManager = Contextx.wifiAwareManager(this)
+inline fun  Context.wifiManagerOrNull(): WifiManager? = Contextx.wifiManagerOrNull(this)
 
 @RequiresApi(api = Build.VERSION_CODES.O)
-fun Context.wifiAwareManagerOrNull(): WifiAwareManager? = Contextx.wifiAwareManagerOrNull(this)
+inline fun  Context.wifiAwareManager(): WifiAwareManager = Contextx.wifiAwareManager(this)
 
-fun Context.wifiP2pManager(): WifiP2pManager = Contextx.wifiP2pManager(this)
+@RequiresApi(api = Build.VERSION_CODES.O)
+inline fun  Context.wifiAwareManagerOrNull(): WifiAwareManager? = Contextx.wifiAwareManagerOrNull(this)
+
+inline fun  Context.wifiP2pManager(): WifiP2pManager = Contextx.wifiP2pManager(this)
 
 // WIFI_SCANNING_SERVICE
 
 @RequiresApi(api = Build.VERSION_CODES.P)
-fun Context.wifiRttManager(): WifiRttManager = Contextx.wifiRttManager(this)
+inline fun  Context.wifiRttManager(): WifiRttManager = Contextx.wifiRttManager(this)
 
-fun Context.nsdManager(): NsdManager = Contextx.nsdManager(this)
+inline fun  Context.nsdManager(): NsdManager = Contextx.nsdManager(this)
 
-fun Context.audioManager(): AudioManager = Contextx.audioManager(this)
-
-@RequiresApi(api = Build.VERSION_CODES.M)
-fun Context.fingerprintManager(): FingerprintManager = Contextx.fingerprintManager(this)
+inline fun  Context.audioManager(): AudioManager = Contextx.audioManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.M)
-fun Context.fingerprintManagerOrNull(): FingerprintManager? = Contextx.fingerprintManagerOrNull(this)
+inline fun  Context.fingerprintManager(): FingerprintManager = Contextx.fingerprintManager(this)
 
-fun Context.mediaRouter(): MediaRouter = Contextx.mediaRouter(this)
+@RequiresApi(api = Build.VERSION_CODES.M)
+inline fun  Context.fingerprintManagerOrNull(): FingerprintManager? = Contextx.fingerprintManagerOrNull(this)
 
-fun Context.telephonyManager(): TelephonyManager = Contextx.telephonyManager(this)
+inline fun  Context.mediaRouter(): MediaRouter = Contextx.mediaRouter(this)
 
-fun Context.telephonyManagerOrNull(): TelephonyManager? = Contextx.telephonyManagerOrNull(this)
+inline fun  Context.telephonyManager(): TelephonyManager = Contextx.telephonyManager(this)
+
+inline fun  Context.telephonyManagerOrNull(): TelephonyManager? = Contextx.telephonyManagerOrNull(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP_MR1)
-fun Context.telephonySubscriptionManager(): SubscriptionManager = Contextx.telephonySubscriptionManager(this)
+inline fun  Context.telephonySubscriptionManager(): SubscriptionManager = Contextx.telephonySubscriptionManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.M)
-fun Context.carrierConfigManager(): CarrierConfigManager = Contextx.carrierConfigManager(this)
+inline fun  Context.carrierConfigManager(): CarrierConfigManager = Contextx.carrierConfigManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.telecomManager(): TelecomManager = Contextx.telecomManager(this)
+inline fun  Context.telecomManager(): TelecomManager = Contextx.telecomManager(this)
 
-fun Context.clipboardManager(): ClipboardManager = Contextx.clipboardManager(this)
+inline fun  Context.clipboardManager(): ClipboardManager = Contextx.clipboardManager(this)
 
-fun Context.inputMethodManager(): InputMethodManager = Contextx.inputMethodManager(this)
+inline fun  Context.inputMethodManager(): InputMethodManager = Contextx.inputMethodManager(this)
 
-fun Context.textServicesManager(): TextServicesManager = Contextx.textServicesManager(this)
+inline fun  Context.textServicesManager(): TextServicesManager = Contextx.textServicesManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.O)
-fun Context.textClassificationManager(): TextClassificationManager = Contextx.textClassificationManager(this)
+inline fun  Context.textClassificationManager(): TextClassificationManager = Contextx.textClassificationManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.appWidgetManager(): AppWidgetManager = Contextx.appWidgetManager(this)
+inline fun  Context.appWidgetManager(): AppWidgetManager = Contextx.appWidgetManager(this)
 
-fun Context.dropBoxManager(): DropBoxManager = Contextx.dropBoxManager(this)
+inline fun  Context.dropBoxManager(): DropBoxManager = Contextx.dropBoxManager(this)
 
-fun Context.devicePolicyManager(): DevicePolicyManager = Contextx.devicePolicyManager(this)
+inline fun  Context.devicePolicyManager(): DevicePolicyManager = Contextx.devicePolicyManager(this)
 
-fun Context.uiModeManager(): UiModeManager = Contextx.uiModeManager(this)
+inline fun  Context.uiModeManager(): UiModeManager = Contextx.uiModeManager(this)
 
-fun Context.downloadManager(): DownloadManager = Contextx.downloadManager(this)
+inline fun  Context.downloadManager(): DownloadManager = Contextx.downloadManager(this)
 
-fun Context.nfcManager(): NfcManager = Contextx.nfcManager(this)
+inline fun  Context.nfcManager(): NfcManager = Contextx.nfcManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.JELLY_BEAN_MR2)
-fun Context.bluetoothManager(): BluetoothManager = Contextx.bluetoothManager(this)
+inline fun  Context.bluetoothManager(): BluetoothManager = Contextx.bluetoothManager(this)
 
-fun Context.usbManager(): UsbManager = Contextx.usbManager(this)
+inline fun  Context.usbManager(): UsbManager = Contextx.usbManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.launcherApps(): LauncherApps = Contextx.launcherApps(this)
+inline fun  Context.launcherApps(): LauncherApps = Contextx.launcherApps(this)
 
-fun Context.inputManager(): InputManager = Contextx.inputManager(this)
+inline fun  Context.inputManager(): InputManager = Contextx.inputManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.JELLY_BEAN_MR1)
-fun Context.displayManager(): DisplayManager = Contextx.displayManager(this)
+inline fun  Context.displayManager(): DisplayManager = Contextx.displayManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.JELLY_BEAN_MR1)
-fun Context.userManager(): UserManager = Contextx.userManager(this)
+inline fun  Context.userManager(): UserManager = Contextx.userManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.restrictionsManager(): RestrictionsManager = Contextx.restrictionsManager(this)
+inline fun  Context.restrictionsManager(): RestrictionsManager = Contextx.restrictionsManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-fun Context.appOpsManager(): AppOpsManager = Contextx.appOpsManager(this)
+inline fun  Context.appOpsManager(): AppOpsManager = Contextx.appOpsManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.cameraManager(): CameraManager = Contextx.cameraManager(this)
+inline fun  Context.cameraManager(): CameraManager = Contextx.cameraManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-fun Context.printManager(): PrintManager = Contextx.printManager(this)
+inline fun  Context.printManager(): PrintManager = Contextx.printManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-fun Context.consumerIrManager(): ConsumerIrManager = Contextx.consumerIrManager(this)
+inline fun  Context.consumerIrManager(): ConsumerIrManager = Contextx.consumerIrManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.tvInputManager(): TvInputManager = Contextx.tvInputManager(this)
+inline fun  Context.tvInputManager(): TvInputManager = Contextx.tvInputManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.tvInputManagerOrNull(): TvInputManager? = Contextx.tvInputManagerOrNull(this)
+inline fun  Context.tvInputManagerOrNull(): TvInputManager? = Contextx.tvInputManagerOrNull(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP_MR1)
-fun Context.usageStatsManager(): UsageStatsManager = Contextx.usageStatsManager(this)
+inline fun  Context.usageStatsManager(): UsageStatsManager = Contextx.usageStatsManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.mediaSessionManager(): MediaSessionManager = Contextx.mediaSessionManager(this)
+inline fun  Context.mediaSessionManager(): MediaSessionManager = Contextx.mediaSessionManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.batteryManager(): BatteryManager = Contextx.batteryManager(this)
+inline fun  Context.batteryManager(): BatteryManager = Contextx.batteryManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.jobScheduler(): JobScheduler = Contextx.jobScheduler(this)
+inline fun  Context.jobScheduler(): JobScheduler = Contextx.jobScheduler(this)
 
 @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-fun Context.mediaProjectionManager(): MediaProjectionManager = Contextx.mediaProjectionManager(this)
+inline fun  Context.mediaProjectionManager(): MediaProjectionManager = Contextx.mediaProjectionManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.M)
-fun Context.midiManager(): MidiManager = Contextx.midiManager(this)
+inline fun  Context.midiManager(): MidiManager = Contextx.midiManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.M)
-fun Context.midiManagerOrNull(): MidiManager? = Contextx.midiManagerOrNull(this)
+inline fun  Context.midiManagerOrNull(): MidiManager? = Contextx.midiManagerOrNull(this)
 
 // RADIO_SERVICE
 
 @RequiresApi(api = Build.VERSION_CODES.N)
-fun Context.hardwarePropertiesManager(): HardwarePropertiesManager = Contextx.hardwarePropertiesManager(this)
+inline fun  Context.hardwarePropertiesManager(): HardwarePropertiesManager = Contextx.hardwarePropertiesManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.N_MR1)
-fun Context.shortcutManager(): ShortcutManager = Contextx.shortcutManager(this)
+inline fun  Context.shortcutManager(): ShortcutManager = Contextx.shortcutManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.N)
-fun Context.systemHealthManager(): SystemHealthManager = Contextx.systemHealthManager(this)
+inline fun  Context.systemHealthManager(): SystemHealthManager = Contextx.systemHealthManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.O)
-fun Context.companionDeviceManager(): CompanionDeviceManager = Contextx.companionDeviceManager(this)
+inline fun  Context.companionDeviceManager(): CompanionDeviceManager = Contextx.companionDeviceManager(this)
 
 @RequiresApi(api = Build.VERSION_CODES.P)
-fun Context.crossProfileApps(): CrossProfileApps = Contextx.crossProfileApps(this)
+inline fun  Context.crossProfileApps(): CrossProfileApps = Contextx.crossProfileApps(this)