+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * :fire: Jsonx.kt format method rename to formatJson
 * :zap: Networkx getWifiState and setWifiEnabled method compatible WifiManager is null
 * :sparkles: Networkx adds getGateway, getDNS1, getDNS2 method
+* :art: Contextx.kt and Viewx.kt partial method changed to inline
 
 
 ## v0.3.2